# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: UPS Gerencia SOA API
  version: 1.0.0
  description: API del sistema distribuido web para solicitudes de recursos academicos.
servers:
  - url: http://localhost:8080
    description: api-gateway
paths:
...
  /content/search:
    get:
      summary: Busca recursos en NASA y NYTimes.
      servers:
        - url: http://localhost:8082
      parameters:
        - name: topic
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recursos consolidados.
...